+++ src/idc/class_lister.py
@@ -5,6 +5,7 @@ def list_classes() -> Dict[str, List[str]]:
     return {
         "seppl.io.Reader": [
             "idc.reader",
+            "idc.reader.depth",
             "idc.reader.imgcls",
             "idc.reader.imgseg",
             "idc.reader.objdet",
@@ -18,6 +19,7 @@ def list_classes() -> Dict[str, List[str]]:
         ],
         "seppl.io.Writer": [
             "idc.writer",
+            "idc.writer.depth",
             "idc.writer.imgcls",
             "idc.writer.imgseg",
             "idc.writer.objdet",