+++ src/idc/reader/depth/__init__.py
@@ -1,0 +1,2 @@
+from ._grayscale import GrayscaleDepthInfoReader
+from ._numpy import NumpyDepthInfoReader