+++ src/idc/writer/depth/__init__.py
@@ -1,0 +1,2 @@
+from ._grayscale import GrayscaleDepthInfoWriter
+from ._numpy import NumpyDepthInfoWriter